+++ CHANGELOG.md
@@ -238,9 +238,11 @@ implemented in the 'jackpolynomials' package, are represented by
 * Function `monomial` to construct monomials like `x_1^4.x_3^7`.
 
 
-## 0.4.0.1 - 2024-XX-XX
+## 0.4.1.0 - 2024-XX-XX
 
-* Fix in `collinearSprays`.
+* There was an error in `gcdSpray`.
+
+* There was a small mistake in `collinearSprays`.
 
 * The functions `changeParameters` and `changeVariables` did not remove the 
 possibly null terms of their result.